<|im_start|>response To update the `go.mod` file, I added new package versions for each modification noted in the Git diff. These updates include adding version `24.0` for `cel.dev/expr`, `121.0` for `cloud.google.com/go` along with its respective `auth` and `longrunning` subpackages, `24.2` for `cloud.google.com/go/monitoring`, `54.0` for `cloud.google.com/go/storage`, and `2ac532fd4443` for the commit hash of `github.com/cncf/xds/go` package.
Reasons for these changes could have been to update dependencies, introduce new features or improvements in the packages. The specific reasons would depend on the context of development practices, such as moving towards a more recent version for better compatibility, security features, or performance enhancements. The commit message succinctly captures these updates and the nature of operation performed on the `go.mod` file.